feat: move portals utils to src/utils
